Refactor iterate_linking_contig_ends
